Refactor and integrate system components for enhanced compatibility
