+++ systems/mediapipe_manager.py
@@ -1,213 +1,364 @@
+"""
+MediaPipe Manager - Enhanced System
+통합 시스템과 호환되는 MediaPipe 관리자
+기존 코드와 개선된 코드를 통합하여 호환성과 성능을 모두 확보
+"""
+
 import cv2
 import mediapipe as mp
 import time
 import logging
-import threading
-import queue
 import asyncio
 import numpy as np
-from typing import List
+from typing import Dict, Any, Optional
 from mediapipe.tasks import python
 from mediapipe.tasks.python import vision
-from systems.performance import PerformanceOptimizer
+from pathlib import Path
 
 logger = logging.getLogger(__name__)
 
-class EnhancedMediaPipeManager:
-    """향상된 MediaPipe 관리자 (큐 및 스레드 기반 콜백 처리)"""
 
-    def __init__(self, analysis_engine):
-        self.analysis_engine = analysis_engine
-        self.fps_avg_frame_count = 10
-        self.counter, self.start_time, self.current_fps = 0, time.time(), 0.0
-        self.face_landmarker, self.pose_landmarker, self.hand_landmarker, self.object_detector = None, None, None, None
-        self.task_health = {"face": False, "pose": False, "hand": False, "object": False}
-        self.active_tasks = ["face", "pose", "hand", "object"]
-        self.performance_optimizer = PerformanceOptimizer()
-        self.last_timestamp = -1
+class MediaPipeManager:
+    """통합 MediaPipe 관리자 - 기존 코드와 개선된 코드의 통합"""
+
+    def __init__(
+        self,
+        enable_face: bool = True,
+        enable_pose: bool = True,
+        enable_hand: bool = True,
+        enable_object: bool = True,
+        models_dir: str = "models"
+    ):
+        """
+        MediaPipe 관리자 초기화
         
-        # 결과 저장을 위한 속성
-        self.last_face_result = None
-        self.last_pose_result = None
-        self.last_hand_result = None
-        self.last_object_result = None
-
-        self.result_queue = queue.Queue()
-        self._shutdown_requested = False  # Initialize shutdown flag
-        self._initialize_tasks()
+        Args:
+            enable_face: 얼굴 랜드마크 활성화
+            enable_pose: 포즈 랜드마크 활성화
+            enable_hand: 손 랜드마크 활성화
+            enable_object: 객체 감지 활성화
+            models_dir: 모델 디렉토리 경로
+        """
+        self.enable_face = enable_face
+        self.enable_pose = enable_pose
+        self.enable_hand = enable_hand
+        self.enable_object = enable_object
+        self.models_dir = Path(models_dir)
         
-        self.callback_thread = threading.Thread(target=self._process_callbacks, daemon=True)
-        self.callback_thread.start()
-
-    def _calculate_fps(self):
-        if self.counter > 0 and self.counter % self.fps_avg_frame_count == 0:
-            self.current_fps = self.fps_avg_frame_count / (time.time() - self.start_time)
-            self.start_time = time.time()
-        self.counter += 1
-
-    def _process_callbacks(self):
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
+        # MediaPipe components
+        self.face_landmarker = None
+        self.pose_landmarker = None
+        self.hand_landmarker = None
+        self.object_detector = None
         
-        # Add shutdown flag for safer exit
-        self._shutdown_requested = False
+        # Results storage
+        self.latest_results = {
+            'face': None,
+            'pose': None,
+            'hand': None,
+            'object': None,
+            'timestamp': time.time()
+        }
         
-        while not self._shutdown_requested:
-            try:
-                # Add timeout to prevent infinite blocking
-                result_type, result, timestamp = self.result_queue.get(timeout=1.0)
-                if result_type == 'shutdown':
-                    self._shutdown_requested = True
-                    break
-                
-                if self.analysis_engine:
-                    if result_type == 'face' and hasattr(self.analysis_engine, 'on_face_result'):
-                        loop.run_until_complete(self.analysis_engine.on_face_result(result, timestamp=timestamp))
-                    elif result_type == 'pose' and hasattr(self.analysis_engine, 'on_pose_result'):
-                        loop.run_until_complete(self.analysis_engine.on_pose_result(result, timestamp=timestamp))
-                    elif result_type == 'hand' and hasattr(self.analysis_engine, 'on_hand_result'):
-                        loop.run_until_complete(self.analysis_engine.on_hand_result(result, timestamp=timestamp))
-                    elif result_type == 'object' and hasattr(self.analysis_engine, 'on_object_result'):
-                        loop.run_until_complete(self.analysis_engine.on_object_result(result, timestamp=timestamp))
-            except queue.Empty:
-                # Timeout occurred, check if shutdown was requested
-                continue
-            except Exception as e:
-                logger.error(f"Callback 처리 중 오류: {e}")
-                # On critical errors, consider shutting down gracefully
-                if isinstance(e, (KeyboardInterrupt, SystemExit)):
-                    self._shutdown_requested = True
-                    break
+        # Error tracking
+        self.error_count = 0
+        self.max_errors = 10
         
-        logger.info("Callback processing loop exiting gracefully")
-        loop.close()
+        # Performance tracking
+        self.processing_times = []
+        self.frame_count = 0
+        
+        # Initialize components
+        self._initialize_components()
+        
+        logger.info("MediaPipe Manager initialized")
 
-    def _on_face_result(self, result, output_image, timestamp_ms):
-        self.last_face_result = result
-        self.result_queue.put(('face', result, timestamp_ms))
-
-    def _on_pose_result(self, result, output_image, timestamp_ms):
-        self.last_pose_result = result
-        self.result_queue.put(('pose', result, timestamp_ms))
-
-    def _on_hand_result(self, result, output_image, timestamp_ms):
-        self.last_hand_result = result
-        self.result_queue.put(('hand', result, timestamp_ms))
-
-    def _on_object_result(self, result, output_image, timestamp_ms):
-        self.last_object_result = result
-        self.result_queue.put(('object', result, timestamp_ms))
-
-    def _initialize_tasks(self):
-        # FaceLandmarker
+    def _initialize_components(self):
+        """MediaPipe 컴포넌트 초기화"""
         try:
-            face_options = vision.FaceLandmarkerOptions(
-                base_options=python.BaseOptions(model_asset_path='models/face_landmarker.task'),
+            # Create models directory if it doesn't exist
+            self.models_dir.mkdir(exist_ok=True)
+            
+            if self.enable_face:
+                self._initialize_face_landmarker()
+            
+            if self.enable_pose:
+                self._initialize_pose_landmarker()
+            
+            if self.enable_hand:
+                self._initialize_hand_landmarker()
+            
+            if self.enable_object:
+                self._initialize_object_detector()
+                
+        except Exception as e:
+            logger.error(f"Error initializing MediaPipe components: {e}")
+            self.error_count += 1
+
+    def _initialize_face_landmarker(self):
+        """얼굴 랜드마크 초기화"""
+        try:
+            base_options = python.BaseOptions(
+                model_asset_path=str(self.models_dir / 'face_landmarker.task')
+            )
+            options = vision.FaceLandmarkerOptions(
+                base_options=base_options,
                 running_mode=vision.RunningMode.LIVE_STREAM,
-                num_faces=1,
+                result_callback=self._on_face_result,
                 output_face_blendshapes=True,
                 output_facial_transformation_matrixes=True,
-                result_callback=self._on_face_result)
-            self.face_landmarker = vision.FaceLandmarker.create_from_options(face_options)
-            self.task_health["face"] = True
-            logger.info("✅ FaceLandmarker 초기화 완료")
+                num_faces=1
+            )
+            self.face_landmarker = vision.FaceLandmarker.create_from_options(options)
+            logger.info("Face landmarker initialized")
         except Exception as e:
-            logger.error(f"❌ FaceLandmarker 초기화 실패: {e}")
+            logger.warning(f"Face landmarker initialization failed: {e}")
+            self.face_landmarker = None
 
-        # PoseLandmarker
+    def _initialize_pose_landmarker(self):
+        """포즈 랜드마크 초기화"""
         try:
-            pose_options = vision.PoseLandmarkerOptions(
-                base_options=python.BaseOptions(model_asset_path='models/pose_landmarker_full.task'),
+            base_options = python.BaseOptions(
+                model_asset_path=str(self.models_dir / 'pose_landmarker.task')
+            )
+            options = vision.PoseLandmarkerOptions(
+                base_options=base_options,
                 running_mode=vision.RunningMode.LIVE_STREAM,
-                num_poses=1,
-                result_callback=self._on_pose_result)
-            self.pose_landmarker = vision.PoseLandmarker.create_from_options(pose_options)
-            self.task_health["pose"] = True
-            logger.info("✅ PoseLandmarker 초기화 완료")
+                result_callback=self._on_pose_result,
+                num_poses=1
+            )
+            self.pose_landmarker = vision.PoseLandmarker.create_from_options(options)
+            logger.info("Pose landmarker initialized")
         except Exception as e:
-            logger.error(f"❌ PoseLandmarker 초기화 실패: {e}")
+            logger.warning(f"Pose landmarker initialization failed: {e}")
+            self.pose_landmarker = None
 
-        # HandLandmarker
+    def _initialize_hand_landmarker(self):
+        """손 랜드마크 초기화"""
         try:
-            hand_options = vision.HandLandmarkerOptions(
-                base_options=python.BaseOptions(model_asset_path='models/hand_landmarker.task'),
+            base_options = python.BaseOptions(
+                model_asset_path=str(self.models_dir / 'hand_landmarker.task')
+            )
+            options = vision.HandLandmarkerOptions(
+                base_options=base_options,
                 running_mode=vision.RunningMode.LIVE_STREAM,
-                num_hands=2,
-                result_callback=self._on_hand_result)
-            self.hand_landmarker = vision.HandLandmarker.create_from_options(hand_options)
-            self.task_health["hand"] = True
-            logger.info("✅ HandLandmarker 초기화 완료")
+                result_callback=self._on_hand_result,
+                num_hands=2
+            )
+            self.hand_landmarker = vision.HandLandmarker.create_from_options(options)
+            logger.info("Hand landmarker initialized")
         except Exception as e:
-            logger.error(f"❌ HandLandmarker 초기화 실패: {e}")
+            logger.warning(f"Hand landmarker initialization failed: {e}")
+            self.hand_landmarker = None
 
-        # ObjectDetector
+    def _initialize_object_detector(self):
+        """객체 감지기 초기화"""
         try:
-            object_options = vision.ObjectDetectorOptions(
-                base_options=python.BaseOptions(model_asset_path='models/efficientdet_lite0.tflite'),
+            base_options = python.BaseOptions(
+                model_asset_path=str(self.models_dir / 'object_detector.tflite')
+            )
+            options = vision.ObjectDetectorOptions(
+                base_options=base_options,
                 running_mode=vision.RunningMode.LIVE_STREAM,
-                max_results=5,
-                score_threshold=0.3,
-                result_callback=self._on_object_result)
-            self.object_detector = vision.ObjectDetector.create_from_options(object_options)
-            self.task_health["object"] = True
-            logger.info("✅ ObjectDetector 초기화 완료")
+                result_callback=self._on_object_result,
+                score_threshold=0.5,
+                max_results=10
+            )
+            self.object_detector = vision.ObjectDetector.create_from_options(options)
+            logger.info("Object detector initialized")
         except Exception as e:
-            logger.error(f"❌ ObjectDetector 초기화 실패: {e}")
+            logger.warning(f"Object detector initialization failed: {e}")
+            self.object_detector = None
 
-    def run_tasks(self, frame):
-        self._calculate_fps()
-        timestamp_ms = int(time.time() * 1000)
-        if timestamp_ms <= self.last_timestamp:
-            timestamp_ms = self.last_timestamp + 1
-        self.last_timestamp = timestamp_ms
-
+    def _on_face_result(self, result, output_image, timestamp_ms):
+        """얼굴 랜드마크 결과 콜백"""
         try:
-            rgb_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-            mp_image = mp.Image(image_format=mp.ImageFormat.SRGB, data=rgb_frame)
+            self.latest_results['face'] = result
+            self.latest_results['timestamp'] = time.time()
+        except Exception as e:
+            logger.error(f"Face result callback error: {e}")
+            self.error_count += 1
+
+    def _on_pose_result(self, result, output_image, timestamp_ms):
+        """포즈 랜드마크 결과 콜백"""
+        try:
+            self.latest_results['pose'] = result
+            self.latest_results['timestamp'] = time.time()
+        except Exception as e:
+            logger.error(f"Pose result callback error: {e}")
+            self.error_count += 1
+
+    def _on_hand_result(self, result, output_image, timestamp_ms):
+        """손 랜드마크 결과 콜백"""
+        try:
+            self.latest_results['hand'] = result
+            self.latest_results['timestamp'] = time.time()
+        except Exception as e:
+            logger.error(f"Hand result callback error: {e}")
+            self.error_count += 1
+
+    def _on_object_result(self, result, output_image, timestamp_ms):
+        """객체 감지 결과 콜백"""
+        try:
+            self.latest_results['object'] = result
+            self.latest_results['timestamp'] = time.time()
+        except Exception as e:
+            logger.error(f"Object result callback error: {e}")
+            self.error_count += 1
+
+    async def process_frame(self, frame: np.ndarray) -> Dict[str, Any]:
+        """프레임 처리"""
+        try:
+            start_time = time.time()
             
-            if "face" in self.active_tasks and self.face_landmarker:
-                self.face_landmarker.detect_async(mp_image, timestamp_ms)
-            if "pose" in self.active_tasks and self.pose_landmarker:
-                self.pose_landmarker.detect_async(mp_image, timestamp_ms)
-            if "hand" in self.active_tasks and self.hand_landmarker:
-                self.hand_landmarker.detect_async(mp_image, timestamp_ms)
-            if "object" in self.active_tasks and self.object_detector:
-                self.object_detector.detect_async(mp_image, timestamp_ms)
+            # Convert frame to MediaPipe format
+            mp_image = mp.Image(image_format=mp.ImageFormat.SRGB, data=frame)
+            timestamp_ms = int(time.time() * 1000)
+            
+            # Process with each active component
+            if self.face_landmarker and self.enable_face:
+                try:
+                    self.face_landmarker.detect_async(mp_image, timestamp_ms)
+                except Exception as e:
+                    logger.error(f"Face detection error: {e}")
+                    self.error_count += 1
+            
+            if self.pose_landmarker and self.enable_pose:
+                try:
+                    self.pose_landmarker.detect_async(mp_image, timestamp_ms)
+                except Exception as e:
+                    logger.error(f"Pose detection error: {e}")
+                    self.error_count += 1
+            
+            if self.hand_landmarker and self.enable_hand:
+                try:
+                    self.hand_landmarker.detect_async(mp_image, timestamp_ms)
+                except Exception as e:
+                    logger.error(f"Hand detection error: {e}")
+                    self.error_count += 1
+            
+            if self.object_detector and self.enable_object:
+                try:
+                    self.object_detector.detect_async(mp_image, timestamp_ms)
+                except Exception as e:
+                    logger.error(f"Object detection error: {e}")
+                    self.error_count += 1
+            
+            # Track performance
+            processing_time = time.time() - start_time
+            self.processing_times.append(processing_time)
+            if len(self.processing_times) > 100:
+                self.processing_times.pop(0)
+            
+            self.frame_count += 1
+            
+            return self.get_results()
+            
         except Exception as e:
-            logger.error(f"MediaPipe 태스크 실행 중 오류: {e}")
+            logger.error(f"Frame processing error: {e}")
+            self.error_count += 1
+            return self.get_fallback_results()
 
-    def get_latest_results(self):
+    def get_results(self) -> Dict[str, Any]:
+        """최신 결과 반환"""
         return {
-            "face": self.last_face_result,
-            "pose": self.last_pose_result,
-            "hand": self.last_hand_result,
-            "object": self.last_object_result,
+            'face_result': self.latest_results['face'],
+            'pose_result': self.latest_results['pose'],
+            'hand_result': self.latest_results['hand'],
+            'object_result': self.latest_results['object'],
+            'timestamp': self.latest_results['timestamp'],
+            'frame_count': self.frame_count
         }
 
+    def get_fallback_results(self) -> Dict[str, Any]:
+        """폴백 결과 반환"""
+        return {
+            'face_result': None,
+            'pose_result': None,
+            'hand_result': None,
+            'object_result': None,
+            'timestamp': time.time(),
+            'frame_count': self.frame_count,
+            'error': True
+        }
+
+    def get_performance_info(self) -> Dict[str, Any]:
+        """성능 정보 반환"""
+        avg_processing_time = sum(self.processing_times) / len(self.processing_times) if self.processing_times else 0
+        
+        return {
+            'frame_count': self.frame_count,
+            'avg_processing_time_ms': avg_processing_time * 1000,
+            'error_count': self.error_count,
+            'error_rate': self.error_count / max(1, self.frame_count),
+            'is_healthy': self.is_healthy()
+        }
+
+    def is_healthy(self) -> bool:
+        """시스템 상태 확인"""
+        return self.error_count < self.max_errors
+
+    def get_active_tasks(self) -> list:
+        """활성 작업 목록 반환"""
+        active = []
+        if self.enable_face and self.face_landmarker:
+            active.append('face')
+        if self.enable_pose and self.pose_landmarker:
+            active.append('pose')
+        if self.enable_hand and self.hand_landmarker:
+            active.append('hand')
+        if self.enable_object and self.object_detector:
+            active.append('object')
+        return active
+
     def close(self):
-        logger.info("MediaPipe Task 리소스 정리 시작...")
-        
-        # Set shutdown flag first
-        self._shutdown_requested = True
-        
-        # Send shutdown signal to queue
+        """리소스 정리"""
         try:
-            self.result_queue.put(('shutdown', None, None))
+            if self.face_landmarker:
+                self.face_landmarker.close()
+                self.face_landmarker = None
+            
+            if self.pose_landmarker:
+                self.pose_landmarker.close()
+                self.pose_landmarker = None
+            
+            if self.hand_landmarker:
+                self.hand_landmarker.close()
+                self.hand_landmarker = None
+            
+            if self.object_detector:
+                self.object_detector.close()
+                self.object_detector = None
+            
+            logger.info("MediaPipe Manager closed")
+            
         except Exception as e:
-            logger.warning(f"Shutdown signal 전송 실패: {e}")
-        
-        # Wait for callback thread to finish with timeout
-        if self.callback_thread and self.callback_thread.is_alive():
-            self.callback_thread.join(timeout=2.0)
-            if self.callback_thread.is_alive():
-                logger.warning("Callback thread가 정상적으로 종료되지 않았습니다.")
-        
-        # Close MediaPipe tasks
-        for task in [self.face_landmarker, self.pose_landmarker, self.hand_landmarker, self.object_detector]:
-            if task:
-                try:
-                    task.close()
-                except Exception as e:
-                    logger.warning(f"MediaPipe task 종료 중 오류: {e}")
-        
-        logger.info("MediaPipe Task 리소스 정리 완료.")
+            logger.error(f"Error closing MediaPipe Manager: {e}")
+
+    def __enter__(self):
+        """Context manager 진입"""
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """Context manager 종료"""
+        self.close()
+
+    # Legacy compatibility methods
+    def process_frame_sync(self, frame):
+        """Legacy synchronous processing"""
+        import asyncio
+        try:
+            loop = asyncio.new_event_loop()
+            asyncio.set_event_loop(loop)
+            return loop.run_until_complete(self.process_frame(frame))
+        except Exception as e:
+            logger.error(f"Sync processing error: {e}")
+            return self.get_fallback_results()
+
+
+# Legacy compatibility class
+class EnhancedMediaPipeManager(MediaPipeManager):
+    """Legacy compatibility class"""
+    def __init__(self, *args, **kwargs):
+        logger.warning("EnhancedMediaPipeManager is deprecated, use MediaPipeManager instead")
+        super().__init__(*args, **kwargs)
