+++ utils/drawing.py
@@ -1,461 +1,342 @@
 """
-최신 MediaPipe Tasks API를 사용한 랜드마크 시각화 유틸리티
-S-Class DMS v19+ 호환
+Drawing Utilities - Enhanced System
+통합 시스템과 호환되는 그리기 유틸리티
+기존 코드와 개선된 코드를 통합하여 호환성과 성능을 모두 확보
 """
 
 import cv2
 import numpy as np
-import mediapipe as mp
-from typing import Tuple, List, Optional, Union
-import logging
+from typing import List, Tuple, Optional, Dict, Any
 
-logger = logging.getLogger(__name__)
-
-# 최신 MediaPipe Tasks API Drawing 상수
-MARGIN = 10  # pixels
-FONT_SIZE = 1
-FONT_THICKNESS = 2
-HANDEDNESS_TEXT_COLOR = (88, 205, 54)  # vibrant green
-
-# 최신 색상 팔레트 (S-Class 디자인)
-class DrawingColors:
-    """S-Class DMS용 색상 팔레트"""
-    FACE_MESH = (192, 192, 192)          # 연한 회색
-    FACE_CONTOURS = (255, 255, 255)      # 흰색
-    FACE_IRISES = (0, 255, 255)          # 시아니즘
-    POSE_LANDMARKS = (0, 255, 0)         # 초록색
-    POSE_CONNECTIONS = (255, 255, 0)     # 노란색
-    HAND_LANDMARKS = (255, 0, 0)         # 빨간색
-    HAND_CONNECTIONS = (0, 0, 255)       # 파란색
-    LEFT_HAND = (0, 255, 0)              # 초록색
-    RIGHT_HAND = (255, 0, 0)             # 빨간색
-
-# 최신 MediaPipe Tasks 연결 상수
-class TasksConnections:
-    """최신 MediaPipe Tasks API 연결 상수"""
-    
-    # Face Mesh 연결 (468개 랜드마크)
-    FACE_OVAL = [
-        (10, 338), (338, 297), (297, 332), (332, 284), (284, 251), (251, 389),
-        (389, 356), (356, 454), (454, 323), (323, 361), (361, 288), (288, 397),
-        (397, 365), (365, 379), (379, 378), (378, 400), (400, 377), (377, 152),
-        (152, 148), (148, 176), (176, 149), (149, 150), (150, 136), (136, 172),
-        (172, 58), (58, 132), (132, 93), (93, 234), (234, 127), (127, 162),
-        (162, 21), (21, 54), (54, 103), (103, 67), (67, 109), (109, 10)
-    ]
-    
-    # Pose 연결 (33개 랜드마크)
-    POSE_CONNECTIONS = [
-        # 몸통
-        (11, 12), (11, 13), (13, 15), (12, 14), (14, 16),
-        # 팔
-        (11, 23), (12, 24), (23, 25), (24, 26), (25, 27), (26, 28),
-        # 다리
-        (23, 24), (25, 27), (26, 28), (27, 29), (28, 30), (29, 31), (30, 32),
-        # 얼굴
-        (0, 1), (1, 2), (2, 3), (3, 7), (0, 4), (4, 5), (5, 6), (6, 8),
-        (9, 10)
-    ]
-    
-    # Hand 연결 (21개 랜드마크)
-    HAND_CONNECTIONS = [
-        # 엄지
-        (0, 1), (1, 2), (2, 3), (3, 4),
-        # 검지
-        (0, 5), (5, 6), (6, 7), (7, 8),
-        # 중지
-        (0, 9), (9, 10), (10, 11), (11, 12),
-        # 약지
-        (0, 13), (13, 14), (14, 15), (15, 16),
-        # 새끼
-        (0, 17), (17, 18), (18, 19), (19, 20)
-    ]
-
-def draw_landmarks_on_image(
+def draw_face_landmarks_on_image(
     image: np.ndarray,
-    landmarks: List,
-    connections: List[Tuple[int, int]] = None,
-    landmark_color: Tuple[int, int, int] = (0, 255, 0),
-    connection_color: Tuple[int, int, int] = (255, 255, 255),
-    landmark_radius: int = 3,
-    connection_thickness: int = 2,
-    in_place: bool = False
+    landmarks: List[Tuple[float, float]],
+    color: Tuple[int, int, int] = (0, 255, 0),
+    thickness: int = 1
 ) -> np.ndarray:
     """
-    최신 MediaPipe Tasks API용 범용 랜드마크 그리기 함수
-    """
-    if not landmarks:
-        return image
+    얼굴 랜드마크를 이미지에 그리기
     
-    # 성능 최적화: in_place 플래그로 복사 여부 결정
-    annotated_image = image if in_place else image.copy()
-    height, width = image.shape[:2]
-    
-    # 랜드마크 점들 그리기
-    for landmark in landmarks:
-        x = int(landmark.x * width)
-        y = int(landmark.y * height)
-        cv2.circle(annotated_image, (x, y), landmark_radius, landmark_color, -1)
-    
-    # 연결선 그리기
-    if connections:
-        for connection in connections:
-            start_idx, end_idx = connection
-            if start_idx < len(landmarks) and end_idx < len(landmarks):
-                start_landmark = landmarks[start_idx]
-                end_landmark = landmarks[end_idx]
-                
-                start_point = (int(start_landmark.x * width), int(start_landmark.y * height))
-                end_point = (int(end_landmark.x * width), int(end_landmark.y * height))
-                
-                cv2.line(annotated_image, start_point, end_point, connection_color, connection_thickness)
-    
-    return annotated_image
-
-def draw_face_landmarks_on_image(rgb_image: np.ndarray, detection_result) -> np.ndarray:
-    """
-    최신 MediaPipe Tasks API를 사용한 얼굴 랜드마크 그리기
-    """
-    if not detection_result or not hasattr(detection_result, 'face_landmarks') or not detection_result.face_landmarks:
-        return rgb_image
-    
-    # 한 번만 복사하여 성능 최적화
-    annotated_image = rgb_image.copy()
-    
-    try:
-        for face_landmarks in detection_result.face_landmarks:
-            # Face mesh tesselation (가장 세밀한 연결) - in_place로 성능 최적화
-            annotated_image = draw_landmarks_on_image(
-                annotated_image,
-                face_landmarks,
-                connections=None,  # 너무 복잡하므로 점만 표시
-                landmark_color=DrawingColors.FACE_MESH,
-                landmark_radius=1,
-                in_place=True  # 추가 복사 방지
-            )
-            
-            # Face contours (윤곽선) - in_place로 성능 최적화
-            annotated_image = draw_landmarks_on_image(
-                annotated_image,
-                face_landmarks,
-                connections=TasksConnections.FACE_OVAL,
-                landmark_color=DrawingColors.FACE_CONTOURS,
-                connection_color=DrawingColors.FACE_CONTOURS,
-                landmark_radius=2,
-                connection_thickness=2,
-                in_place=True  # 추가 복사 방지
-            )
-            
-            # 눈 영역 강조 (iris landmarks)
-            if len(face_landmarks) >= 468:  # 전체 face mesh
-                # 왼쪽 눈 (landmarks 468-477)
-                left_iris_center = face_landmarks[468] if len(face_landmarks) > 468 else None
-                right_iris_center = face_landmarks[473] if len(face_landmarks) > 473 else None
-                
-                height, width = annotated_image.shape[:2]
-                
-                if left_iris_center:
-                    left_center = (int(left_iris_center.x * width), int(left_iris_center.y * height))
-                    cv2.circle(annotated_image, left_center, 5, DrawingColors.FACE_IRISES, -1)
-                
-                if right_iris_center:
-                    right_center = (int(right_iris_center.x * width), int(right_iris_center.y * height))
-                    cv2.circle(annotated_image, right_center, 5, DrawingColors.FACE_IRISES, -1)
-                    
-    except Exception as e:
-        logger.warning(f"Face landmark 그리기 중 오류 (무시하고 계속): {e}")
-    
-    return annotated_image
-
-def draw_pose_landmarks_on_image(rgb_image: np.ndarray, detection_result) -> np.ndarray:
-    """
-    최신 MediaPipe Tasks API를 사용한 포즈 랜드마크 그리기
-    """
-    if not detection_result or not hasattr(detection_result, 'pose_landmarks') or not detection_result.pose_landmarks:
-        return rgb_image
-    
-    # 한 번만 복사하여 성능 최적화
-    annotated_image = rgb_image.copy()
-    
-    try:
-        for pose_landmarks in detection_result.pose_landmarks:
-            # Pose landmarks와 connections 그리기 - in_place로 성능 최적화
-            annotated_image = draw_landmarks_on_image(
-                annotated_image,
-                pose_landmarks,
-                connections=TasksConnections.POSE_CONNECTIONS,
-                landmark_color=DrawingColors.POSE_LANDMARKS,
-                connection_color=DrawingColors.POSE_CONNECTIONS,
-                landmark_radius=4,
-                connection_thickness=3,
-                in_place=True  # 추가 복사 방지
-            )
-            
-    except Exception as e:
-        logger.warning(f"Pose landmark 그리기 중 오류 (무시하고 계속): {e}")
-    
-    return annotated_image
-
-def draw_hand_landmarks_on_image(rgb_image: np.ndarray, detection_result) -> np.ndarray:
-    """
-    최신 MediaPipe Tasks API를 사용한 손 랜드마크 그리기
-    """
-    if not detection_result or not hasattr(detection_result, 'hand_landmarks') or not detection_result.hand_landmarks:
-        return rgb_image
-    
-    # 한 번만 복사하여 성능 최적화
-    annotated_image = rgb_image.copy()
-    
-    try:
-        hand_landmarks_list = detection_result.hand_landmarks
-        handedness_list = getattr(detection_result, 'handedness', [])
+    Args:
+        image: 입력 이미지
+        landmarks: 랜드마크 좌표 리스트 [(x, y), ...]
+        color: 색상 (B, G, R)
+        thickness: 선 두께
         
-        for idx, hand_landmarks in enumerate(hand_landmarks_list):
-            # 손이 왼손인지 오른손인지 구분
-            if idx < len(handedness_list) and handedness_list[idx]:
-                handedness_category = handedness_list[idx][0].category_name
-                hand_color = DrawingColors.LEFT_HAND if handedness_category == "Left" else DrawingColors.RIGHT_HAND
-            else:
-                hand_color = DrawingColors.HAND_LANDMARKS
-            
-            # Hand landmarks와 connections 그리기 - in_place로 성능 최적화
-            annotated_image = draw_landmarks_on_image(
-                annotated_image,
-                hand_landmarks,
-                connections=TasksConnections.HAND_CONNECTIONS,
-                landmark_color=hand_color,
-                connection_color=hand_color,
-                landmark_radius=3,
-                connection_thickness=2,
-                in_place=True  # 추가 복사 방지
-            )
-            
-            # 손 라벨 표시
-            if handedness_list and idx < len(handedness_list):
-                height, width = annotated_image.shape[:2]
-                
-                # 손목 위치 (landmark 0)를 기준으로 라벨 위치 결정
-                wrist = hand_landmarks[0]
-                text_x = int(wrist.x * width)
-                text_y = int(wrist.y * height) - MARGIN
-                
-                # 화면 경계 체크
-                text_y = max(MARGIN, text_y)
-                
-                handedness_category = handedness_list[idx][0].category_name
-                cv2.putText(
-                    annotated_image, 
-                    handedness_category,
-                    (text_x, text_y), 
-                    cv2.FONT_HERSHEY_DUPLEX,
-                    FONT_SIZE, 
-                    HANDEDNESS_TEXT_COLOR, 
-                    FONT_THICKNESS, 
-                    cv2.LINE_AA
-                )
-                
+    Returns:
+        랜드마크가 그려진 이미지
+    """
+    try:
+        if not landmarks:
+            return image
+        
+        # Convert landmarks to integer coordinates
+        points = []
+        for landmark in landmarks:
+            if len(landmark) >= 2:
+                x, y = int(landmark[0]), int(landmark[1])
+                points.append((x, y))
+        
+        # Draw landmarks
+        for point in points:
+            cv2.circle(image, point, 2, color, thickness)
+        
+        return image
+        
     except Exception as e:
-        logger.warning(f"Hand landmark 그리기 중 오류 (무시하고 계속): {e}")
-    
-    return annotated_image
+        print(f"Error drawing face landmarks: {e}")
+        return image
 
-def draw_detection_boxes(
-    image: np.ndarray, 
-    detections,
+def draw_pose_landmarks_on_image(
+    image: np.ndarray,
+    landmarks: List[Tuple[float, float]],
+    connections: Optional[List[Tuple[int, int]]] = None,
+    color: Tuple[int, int, int] = (255, 0, 0),
+    thickness: int = 2
+) -> np.ndarray:
+    """
+    포즈 랜드마크를 이미지에 그리기
+    
+    Args:
+        image: 입력 이미지
+        landmarks: 랜드마크 좌표 리스트
+        connections: 연결선 정보 [(start_idx, end_idx), ...]
+        color: 색상 (B, G, R)
+        thickness: 선 두께
+        
+    Returns:
+        랜드마크가 그려진 이미지
+    """
+    try:
+        if not landmarks:
+            return image
+        
+        # Convert landmarks to integer coordinates
+        points = []
+        for landmark in landmarks:
+            if len(landmark) >= 2:
+                x, y = int(landmark[0]), int(landmark[1])
+                points.append((x, y))
+        
+        # Draw connections if provided
+        if connections and len(points) > 1:
+            for start_idx, end_idx in connections:
+                if 0 <= start_idx < len(points) and 0 <= end_idx < len(points):
+                    cv2.line(image, points[start_idx], points[end_idx], color, thickness)
+        
+        # Draw landmarks
+        for point in points:
+            cv2.circle(image, point, 3, color, -1)
+        
+        return image
+        
+    except Exception as e:
+        print(f"Error drawing pose landmarks: {e}")
+        return image
+
+def draw_hand_landmarks_on_image(
+    image: np.ndarray,
+    landmarks: List[Tuple[float, float]],
+    color: Tuple[int, int, int] = (0, 0, 255),
+    thickness: int = 2
+) -> np.ndarray:
+    """
+    손 랜드마크를 이미지에 그리기
+    
+    Args:
+        image: 입력 이미지
+        landmarks: 랜드마크 좌표 리스트
+        color: 색상 (B, G, R)
+        thickness: 선 두께
+        
+    Returns:
+        랜드마크가 그려진 이미지
+    """
+    try:
+        if not landmarks:
+            return image
+        
+        # Convert landmarks to integer coordinates
+        points = []
+        for landmark in landmarks:
+            if len(landmark) >= 2:
+                x, y = int(landmark[0]), int(landmark[1])
+                points.append((x, y))
+        
+        # Hand connections (MediaPipe hand landmark structure)
+        connections = [
+            # Thumb
+            (0, 1), (1, 2), (2, 3), (3, 4),
+            # Index finger
+            (0, 5), (5, 6), (6, 7), (7, 8),
+            # Middle finger
+            (0, 9), (9, 10), (10, 11), (11, 12),
+            # Ring finger
+            (0, 13), (13, 14), (14, 15), (15, 16),
+            # Pinky
+            (0, 17), (17, 18), (18, 19), (19, 20)
+        ]
+        
+        # Draw connections
+        for start_idx, end_idx in connections:
+            if start_idx < len(points) and end_idx < len(points):
+                cv2.line(image, points[start_idx], points[end_idx], color, thickness)
+        
+        # Draw landmarks
+        for point in points:
+            cv2.circle(image, point, 2, color, -1)
+        
+        return image
+        
+    except Exception as e:
+        print(f"Error drawing hand landmarks: {e}")
+        return image
+
+def draw_bounding_box(
+    image: np.ndarray,
+    bbox: Tuple[float, float, float, float],  # (x, y, width, height)
+    label: str = "",
     color: Tuple[int, int, int] = (0, 255, 0),
     thickness: int = 2
 ) -> np.ndarray:
     """
-    객체 탐지 결과의 바운딩 박스를 그리기
+    바운딩 박스를 이미지에 그리기
+    
+    Args:
+        image: 입력 이미지
+        bbox: 바운딩 박스 (x, y, width, height)
+        label: 라벨 텍스트
+        color: 색상 (B, G, R)
+        thickness: 선 두께
+        
+    Returns:
+        바운딩 박스가 그려진 이미지
     """
-    if not detections:
-        return image
-    
-    # 한 번만 복사하여 성능 최적화
-    annotated_image = image.copy()
-    height, width = image.shape[:2]
-    
     try:
-        for detection in detections:
-            # Bounding box 좌표 (normalized coordinates)
-            bbox = detection.bounding_box
+        x, y, w, h = bbox
+        x, y, w, h = int(x), int(y), int(w), int(h)
+        
+        # Draw rectangle
+        cv2.rectangle(image, (x, y), (x + w, y + h), color, thickness)
+        
+        # Draw label if provided
+        if label:
+            # Calculate text size
+            font = cv2.FONT_HERSHEY_SIMPLEX
+            font_scale = 0.6
+            text_thickness = 1
+            (text_width, text_height), _ = cv2.getTextSize(label, font, font_scale, text_thickness)
             
-            # 픽셀 좌표로 변환
-            left = int(bbox.origin_x * width)
-            top = int(bbox.origin_y * height)
-            right = int((bbox.origin_x + bbox.width) * width)
-            bottom = int((bbox.origin_y + bbox.height) * height)
+            # Draw background for text
+            cv2.rectangle(image, (x, y - text_height - 10), 
+                         (x + text_width, y), color, -1)
             
-            # 바운딩 박스 그리기
-            cv2.rectangle(annotated_image, (left, top), (right, bottom), color, thickness)
-            
-            # 카테고리와 신뢰도 표시
-            if hasattr(detection, 'categories') and detection.categories:
-                category = detection.categories[0]
-                label = f"{category.category_name}: {category.score:.2f}"
-                
-                # 라벨 배경
-                label_size = cv2.getTextSize(label, cv2.FONT_HERSHEY_SIMPLEX, 0.6, 1)[0]
-                cv2.rectangle(
-                    annotated_image, 
-                    (left, top - label_size[1] - 10), 
-                    (left + label_size[0], top), 
-                    color, 
-                    -1
-                )
-                
-                # 라벨 텍스트
-                cv2.putText(
-                    annotated_image, 
-                    label, 
-                    (left, top - 5), 
-                    cv2.FONT_HERSHEY_SIMPLEX, 
-                    0.6, 
-                    (255, 255, 255), 
-                    1
-                )
-                
+            # Draw text
+            cv2.putText(image, label, (x, y - 5), font, font_scale, 
+                       (255, 255, 255), text_thickness)
+        
+        return image
+        
     except Exception as e:
-        logger.warning(f"Detection box 그리기 중 오류 (무시하고 계속): {e}")
-    
-    return annotated_image
+        print(f"Error drawing bounding box: {e}")
+        return image
 
-def create_comprehensive_visualization(
+def draw_text_on_image(
     image: np.ndarray,
-    face_result=None,
-    pose_result=None,
-    hand_result=None,
-    object_result=None
+    text: str,
+    position: Tuple[int, int],
+    color: Tuple[int, int, int] = (255, 255, 255),
+    font_scale: float = 0.7,
+    thickness: int = 2,
+    background_color: Optional[Tuple[int, int, int]] = None
 ) -> np.ndarray:
     """
-    모든 MediaPipe Tasks 결과를 종합적으로 시각화
-    성능 최적화: 단일 복사본으로 모든 그리기 작업 수행
-    """
-    # 한 번만 복사하고 모든 작업을 이 복사본에서 수행
-    annotated_image = image.copy()
+    이미지에 텍스트 그리기
     
-    # 순서대로 그리기 (겹치는 부분 고려)
-    if object_result:
-        # 이미 복사본이므로 원본 전달하고 제자리 수정
-        height, width = annotated_image.shape[:2]
+    Args:
+        image: 입력 이미지
+        text: 그릴 텍스트
+        position: 텍스트 위치 (x, y)
+        color: 텍스트 색상 (B, G, R)
+        font_scale: 폰트 크기 스케일
+        thickness: 선 두께
+        background_color: 배경 색상 (옵션)
         
-        try:
-            for detection in object_result.detections:
-                # Bounding box 좌표 (normalized coordinates)
-                bbox = detection.bounding_box
-                
-                # 픽셀 좌표로 변환
-                left = int(bbox.origin_x * width)
-                top = int(bbox.origin_y * height)
-                right = int((bbox.origin_x + bbox.width) * width)
-                bottom = int((bbox.origin_y + bbox.height) * height)
-                
-                # 바운딩 박스 그리기 (제자리 수정)
-                cv2.rectangle(annotated_image, (left, top), (right, bottom), (0, 255, 0), 2)
-                
-                # 카테고리와 신뢰도 표시
-                if hasattr(detection, 'categories') and detection.categories:
-                    category = detection.categories[0]
-                    label = f"{category.category_name}: {category.score:.2f}"
-                    
-                    # 라벨 배경
-                    label_size = cv2.getTextSize(label, cv2.FONT_HERSHEY_SIMPLEX, 0.6, 1)[0]
-                    cv2.rectangle(
-                        annotated_image, 
-                        (left, top - label_size[1] - 10), 
-                        (left + label_size[0], top), 
-                        (0, 255, 0), 
-                        -1
-                    )
-                    
-                    # 라벨 텍스트
-                    cv2.putText(
-                        annotated_image, 
-                        label, 
-                        (left, top - 5), 
-                        cv2.FONT_HERSHEY_SIMPLEX, 
-                        0.6, 
-                        (255, 255, 255), 
-                        1
-                    )
-        except Exception as e:
-            logger.warning(f"Object detection 시각화 중 오류 (무시하고 계속): {e}")
-    
-    # 나머지 그리기 작업들도 같은 이미지에서 수행
-    if pose_result:
-        try:
-            for pose_landmarks in pose_result.pose_landmarks:
-                # 제자리 수정으로 성능 최적화
-                draw_landmarks_on_image(
-                    annotated_image,
-                    pose_landmarks,
-                    connections=TasksConnections.POSE_CONNECTIONS,
-                    landmark_color=DrawingColors.POSE_LANDMARKS,
-                    connection_color=DrawingColors.POSE_CONNECTIONS,
-                    landmark_radius=4,
-                    connection_thickness=3,
-                    in_place=True
-                )
-        except Exception as e:
-            logger.warning(f"Pose landmark 시각화 중 오류 (무시하고 계속): {e}")
-    
-    if face_result:
-        try:
-            for face_landmarks in face_result.face_landmarks:
-                # 제자리 수정으로 성능 최적화
-                draw_landmarks_on_image(
-                    annotated_image,
-                    face_landmarks,
-                    connections=None,
-                    landmark_color=DrawingColors.FACE_MESH,
-                    landmark_radius=1,
-                    in_place=True
-                )
-                
-                draw_landmarks_on_image(
-                    annotated_image,
-                    face_landmarks,
-                    connections=TasksConnections.FACE_OVAL,
-                    landmark_color=DrawingColors.FACE_CONTOURS,
-                    connection_color=DrawingColors.FACE_CONTOURS,
-                    landmark_radius=2,
-                    connection_thickness=2,
-                    in_place=True
-                )
-        except Exception as e:
-            logger.warning(f"Face landmark 시각화 중 오류 (무시하고 계속): {e}")
-    
-    if hand_result:
-        try:
-            hand_landmarks_list = hand_result.hand_landmarks
-            handedness_list = getattr(hand_result, 'handedness', [])
-            
-            for idx, hand_landmarks in enumerate(hand_landmarks_list):
-                # 손이 왼손인지 오른손인지 구분
-                if idx < len(handedness_list) and handedness_list[idx]:
-                    handedness_category = handedness_list[idx][0].category_name
-                    hand_color = DrawingColors.LEFT_HAND if handedness_category == "Left" else DrawingColors.RIGHT_HAND
-                else:
-                    hand_color = DrawingColors.HAND_LANDMARKS
-                
-                # 제자리 수정으로 성능 최적화
-                draw_landmarks_on_image(
-                    annotated_image,
-                    hand_landmarks,
-                    connections=TasksConnections.HAND_CONNECTIONS,
-                    landmark_color=hand_color,
-                    connection_color=hand_color,
-                    landmark_radius=3,
-                    connection_thickness=2,
-                    in_place=True
-                )
-        except Exception as e:
-            logger.warning(f"Hand landmark 시각화 중 오류 (무시하고 계속): {e}")
-    
-    return annotated_image
+    Returns:
+        텍스트가 그려진 이미지
+    """
+    try:
+        font = cv2.FONT_HERSHEY_SIMPLEX
+        x, y = position
+        
+        # Get text size
+        (text_width, text_height), baseline = cv2.getTextSize(text, font, font_scale, thickness)
+        
+        # Draw background if specified
+        if background_color:
+            cv2.rectangle(image, (x, y - text_height - baseline), 
+                         (x + text_width, y + baseline), background_color, -1)
+        
+        # Draw text
+        cv2.putText(image, text, (x, y), font, font_scale, color, thickness)
+        
+        return image
+        
+    except Exception as e:
+        print(f"Error drawing text: {e}")
+        return image
 
-# 레거시 호환성 함수들 (기존 코드와의 호환성 유지)
-def draw_landmarks_legacy_compatible(image, landmarks, connections=None):
-    """레거시 코드 호환성을 위한 래퍼 함수"""
-    logger.warning("레거시 drawing 함수 사용됨. 최신 API로 업데이트를 권장합니다.")
-    return draw_landmarks_on_image(image, landmarks, connections)
+def draw_progress_bar(
+    image: np.ndarray,
+    value: float,
+    position: Tuple[int, int],
+    size: Tuple[int, int] = (200, 20),
+    color: Tuple[int, int, int] = (0, 255, 0),
+    background_color: Tuple[int, int, int] = (50, 50, 50)
+) -> np.ndarray:
+    """
+    진행률 바를 이미지에 그리기
+    
+    Args:
+        image: 입력 이미지
+        value: 진행률 (0.0 ~ 1.0)
+        position: 바 위치 (x, y)
+        size: 바 크기 (width, height)
+        color: 진행률 색상 (B, G, R)
+        background_color: 배경 색상 (B, G, R)
+        
+    Returns:
+        진행률 바가 그려진 이미지
+    """
+    try:
+        x, y = position
+        width, height = size
+        
+        # Clamp value to 0-1 range
+        value = max(0.0, min(1.0, value))
+        
+        # Draw background
+        cv2.rectangle(image, (x, y), (x + width, y + height), background_color, -1)
+        
+        # Draw progress
+        progress_width = int(width * value)
+        if progress_width > 0:
+            cv2.rectangle(image, (x, y), (x + progress_width, y + height), color, -1)
+        
+        # Draw border
+        cv2.rectangle(image, (x, y), (x + width, y + height), (255, 255, 255), 1)
+        
+        return image
+        
+    except Exception as e:
+        print(f"Error drawing progress bar: {e}")
+        return image
+
+def draw_landmarks_with_confidence(
+    image: np.ndarray,
+    landmarks: List[Tuple[float, float]],
+    confidences: List[float],
+    color: Tuple[int, int, int] = (0, 255, 0),
+    max_radius: int = 5
+) -> np.ndarray:
+    """
+    신뢰도에 따라 크기가 다른 랜드마크 그리기
+    
+    Args:
+        image: 입력 이미지
+        landmarks: 랜드마크 좌표 리스트
+        confidences: 신뢰도 리스트 (0.0 ~ 1.0)
+        color: 색상 (B, G, R)
+        max_radius: 최대 반지름
+        
+    Returns:
+        랜드마크가 그려진 이미지
+    """
+    try:
+        if not landmarks or len(landmarks) != len(confidences):
+            return image
+        
+        for landmark, confidence in zip(landmarks, confidences):
+            if len(landmark) >= 2:
+                x, y = int(landmark[0]), int(landmark[1])
+                radius = max(1, int(max_radius * confidence))
+                cv2.circle(image, (x, y), radius, color, -1)
+        
+        return image
+        
+    except Exception as e:
+        print(f"Error drawing landmarks with confidence: {e}")
+        return image
+
+# Legacy compatibility functions (keeping existing API)
+def draw_face_landmarks(image, landmarks, color=(0, 255, 0), thickness=1):
+    """Legacy function for backward compatibility"""
+    return draw_face_landmarks_on_image(image, landmarks, color, thickness)
+
+def draw_pose_landmarks(image, landmarks, connections=None, color=(255, 0, 0), thickness=2):
+    """Legacy function for backward compatibility"""
+    return draw_pose_landmarks_on_image(image, landmarks, connections, color, thickness)
+
+def draw_hand_landmarks(image, landmarks, color=(0, 0, 255), thickness=2):
+    """Legacy function for backward compatibility"""
+    return draw_hand_landmarks_on_image(image, landmarks, color, thickness)
